Support a Max Queue Size for Buffered Messages
Adds a variable maxEnqueuedMessages that handles setting a limit
on the maximum number of messages to queue up when a connection
is down.
Defaults to Infinity, and can be used to disable queueing by setting
the value to 0.
Added a small test around the new feature, and fixed an oddity in an
existing test.